infra/brain/requirements.txt
fastapi
uvicorn[standard]
google-genai
aiohttp
loguru
